% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterName.R
\name{adjacencyFromSimilarity}
\alias{adjacencyFromSimilarity}
\title{Adjacency Matrix}
\usage{
adjacencyFromSimilarity(m, clusters)
}
\arguments{
\item{m}{similarity matrix}

\item{clusters}{list of clusters}
}
\description{
Creates an adjacency matrix from similarity matrix, connecting nodes that belong to
the same cluster.
}
