% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aPEAR-package.R
\docType{package}
\name{aPEAR-package}
\alias{aPEAR}
\alias{aPEAR-package}
\title{aPEAR: Advanced Pathway Enrichment Analysis Representation}
\description{
Simplify pathway enrichment analysis results by detecting clusters of similar pathways and visualizing it as an enrichment network, where nodes and edges describe the pathways and similarity between them, respectively. This reduces the redundancy of the overlapping pathways and helps to notice the most important biological themes in the data (Kerseviciute and Gordevicius (2023) \doi{10.1101/2023.03.28.534514}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/vugene/aPEAR}
  \item Report bugs at \url{https://gitlab.com/vugene/aPEAR/-/issues}
}

}
\author{
\strong{Maintainer}: Ieva Kerseviciute \email{kerseviciute.ieva@gmail.com} (\href{https://orcid.org/0000-0002-9353-038X}{ORCID})

Other contributors:
\itemize{
  \item Juozas Gordevicius \email{juozas@vugene.com} (\href{https://orcid.org/0000-0003-3190-6721}{ORCID}) [thesis advisor]
  \item VUGENE, LLC [copyright holder, funder]
}

}
\keyword{internal}
