% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{clusterNamesHits}
\alias{clusterNamesHits}
\title{Selects a name for each pathway cluster using hits method.}
\usage{
clusterNamesHits(sim, clusters)
}
\arguments{
\item{sim}{a similarity matrix used to detect the clusters}

\item{clusters}{a vector of clusters, with names indicating the pathway name}
}
\description{
Selects a name for each pathway cluster using hits method.
}
