% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterName.R
\name{clusterNamesPagerank}
\alias{clusterNamesPagerank}
\title{PageRank}
\usage{
clusterNamesPagerank(sim, clusters)
}
\arguments{
\item{sim}{a similarity matrix used to detect the clusters}

\item{clusters}{a vector of clusters, with names indicating the pathway name}
}
\description{
Selects a name for each pathway cluster using PageRank method.
}
