% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{clusterNamesWordcloud}
\alias{clusterNamesWordcloud}
\title{Selects a name for each pathway cluster using word cloud method.
TODO: how to use python in R packages?}
\usage{
clusterNamesWordcloud(clusters)
}
\description{
Selects a name for each pathway cluster using word cloud method.
TODO: how to use python in R packages?
}
