% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{emptyMatrix}
\alias{emptyMatrix}
\title{Empty Matrix}
\usage{
emptyMatrix(values, data = NA)
}
\arguments{
\item{values}{a list of values}

\item{data}{default data}
}
\description{
Creates an empty matrix from names of a list. Will set colnames and rownames as
the names of the list.
}
