% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{enrichmentData}
\alias{enrichmentData}
\title{Creating Valid Input}
\description{
Currently, this package is able to work with DOSE enrichment classes. However, you
may wish to pass different enrichment results. In this case, prepare your data so that it
contains 'Description' and 'core_enrichment' columns. Use \code{'colorBy'} and \code{'nodeSize'}
parameters to set which columns should be used for colouring the nodes and adjusting their size
(eg. \code{colorBy = 'NES', nodeSize = 'setSize'}).
}
\seealso{
\code{enrichmentNetwork}
}
