% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkPlot.R
\name{enrichmentNetwork.addEllipses}
\alias{enrichmentNetwork.addEllipses}
\title{Enrichment Network}
\usage{
enrichmentNetwork.addEllipses(plot, pathways)
}
\arguments{
\item{plot}{\code{ggplot} object}

\item{pathways}{a data table containing pathway coordinates in the graph and their clustering
results}
}
\description{
Adds ellipses around pathway clusters.
}
