% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkPlot.R
\name{enrichmentNetwork.clusterLabels}
\alias{enrichmentNetwork.clusterLabels}
\title{Cluster Labels}
\usage{
enrichmentNetwork.clusterLabels(pathways, fontSize = 5, repelLabels = FALSE)
}
\arguments{
\item{pathways}{a data table containing pathway coordinates in the graph and their clustering
results}

\item{fontSize}{adjust cluster label font size}

\item{repelLabels}{should cluster label positions be corrected}
}
\value{
A \code{geom_text} object with labels and their coordinates, ready to add to a
\code{ggplot} object.
}
\description{
Displays a label for each pathway cluster in the graph.
}
