% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkPlot.R
\name{enrichmentNetwork.connect}
\alias{enrichmentNetwork.connect}
\title{Enrichment Network}
\usage{
enrichmentNetwork.connect(sim, clusters, innerCutoff = 0.1, outerCutoff = 0.5)
}
\arguments{
\item{sim}{similarity matrix}

\item{clusters}{a list of clusters}

\item{innerCutoff}{similarity cutoff for in-cluster nodes}

\item{outerCutoff}{similarity cutoff for between-cluster nodes}
}
\value{
A list with two objects: \code{coordinates} - each pathway coordinates in the graph, and
\code{edges} - list of edges in the graph, with their coordinates.
}
\description{
Connects related pathways.
}
