% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkPlot.R
\name{enrichmentNetwork.ellipse}
\alias{enrichmentNetwork.ellipse}
\title{Enrichment Network}
\usage{
enrichmentNetwork.ellipse(c, a, b, alpha, bigger = 0)
}
\arguments{
\item{c}{ellipse center}

\item{a}{a axis}

\item{b}{b axis}

\item{alpha}{ellipse angle}

\item{bigger}{how much to increase ellipse size}
}
\description{
Calculates ellipse coordinates from its equation.
}
