% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkPlot.R
\name{enrichmentNetwork.plot}
\alias{enrichmentNetwork.plot}
\title{Enrichment Network}
\usage{
enrichmentNetwork.plot(dt, sim, clust, innerCutoff = 0.1, outerCutoff = 0.5)
}
\arguments{
\item{dt}{output from \code{enrichmentNetwork.prepareEnrichmentClusters}}

\item{sim}{similarity matrix}

\item{innerCutoff}{similarity cutoff for in-cluster nodes}

\item{outerCutoff}{similarity cutoff for between-cluster nodes}
}
\value{
\code{ggplot} object.
}
\description{
Creates a \code{ggplot} object from pathway similarity matrix and clustering results.
}
