% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkPlot.R
\name{enrichmentNetwork.prepareEnrichmentClusters}
\alias{enrichmentNetwork.prepareEnrichmentClusters}
\title{Enrichment Clusters}
\usage{
enrichmentNetwork.prepareEnrichmentClusters(enrichment, clusters, params)
}
\description{
Prepares enrichment and clustering results for \code{enrichmentNetwork.plot} method.
}
