% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterName.R
\name{findClusterNames}
\alias{findClusterNames}
\title{Cluster Name}
\usage{
findClusterNames(sim, clusters, method = c("pagerank", "hits", "none"))
}
\arguments{
\item{sim}{similarity matrix used to detect the clusters}

\item{clusters}{a list of clusters, where values are cluster ID and names are pathway names}

\item{method}{a method for setting cluster names. Available method include \code{'pagerank'},
\code{'hits'} and \code{'none'}}
}
\description{
Selects best fitting name for a cluster of pathways.
}
