% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{findClusters}
\alias{findClusters}
\title{Finds clusters in a similarity matrix.}
\usage{
findClusters(
  sim,
  minClusterSize = 2,
  method = "markov",
  nameMethod = "pagerank"
)
}
\arguments{
\item{sim}{similarity matrix}

\item{minClusterSize}{minimum cluster size (integer, >= 1). This parameter is ignored when using
\code{method = 'spectral'}}

\item{method}{a method to be used for detecting clusters. Available options include 'markov',
'hier' and 'spectral'}

\item{nameMethod}{a method for setting cluster names. Available method include 'pagerank',
'hits', 'wordcloud' and 'none'}
}
\description{
Finds clusters in a similarity matrix.
}
