% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{findClustersHier}
\alias{findClustersHier}
\title{Finds clusters using hierarchical algorithm.}
\usage{
findClustersHier(sim, minClusterSize = 2, method = "pagerank")
}
\arguments{
\item{sim}{similarity matrix}

\item{minClusterSize}{minimum cluster size}

\item{method}{a method for setting the cluster names}
}
\description{
Finds clusters using hierarchical algorithm.
}
