% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{findClustersMarkov}
\alias{findClustersMarkov}
\title{Finds clusters in a similarity matrix.}
\usage{
findClustersMarkov(sim, minClusterSize = 2, method = "pagerank")
}
\arguments{
\item{sim}{similarity matrix}

\item{minClusterSize}{minimum cluster size}

\item{method}{a method for setting the cluster names}
}
\description{
Finds clusters in a similarity matrix.
}
