% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{findClustersSpectral}
\alias{findClustersSpectral}
\title{Finds clusters using adaptive spectral clustering.}
\usage{
findClustersSpectral(sim, method = "wordcloud")
}
\arguments{
\item{sim}{similarity matrix}

\item{method}{a method for setting the cluster names}
}
\description{
Finds clusters using adaptive spectral clustering.
}
