% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{getGenes}
\alias{getGenes}
\title{Enrichment Genes}
\usage{
getGenes(enrichment, geneCol, pathCol = "Description")
}
\arguments{
\item{enrichment}{a data frame containing enrichment results}

\item{geneCol}{which column contains gene lists}

\item{pathCol}{which column contains path descriptions (human readable text, not ids!)}
}
\description{
Gets genes involved in each pathway of enrichment result.
}
