% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipse.R
\name{mvee}
\alias{mvee}
\title{Minimum Volume Enclosing Ellipsoid}
\usage{
mvee(
  xy = NULL,
  tolerance = 0.005,
  plotme = TRUE,
  max.iter = 500,
  shiftxy = TRUE
)
}
\arguments{
\item{xy}{a two-column data frame containing x and y coordinates if NULL then a random sample
set of 10 points will be generated}

\item{tolerance}{a tolerance value (default = 0.005)}

\item{plotme}{FALSE/TRUE. Plots the points and ellipse. Default TRUE.}

\item{max.iter}{The maximum number of iterations. If the script tries this number of iterations
but still can't get to the tolerance value, it displays an error message and returns NULL}

\item{shiftxy}{TRUE/FALSE. If True, will apply a shift to the coordinates to make them
smaller and speed up the matrix calculations, then reverse the shift to the center point of the
resulting ellipoid. Default TRUE}
}
\description{
Uses the Khachiyan Algorithm to find the the minimum volume enclosing
ellipsoid (MVEE) of a set of points. In two dimensions, this is just
the bounding ellipse (this function is limited to two dimensions).
Adapted by Andy Lyons from Matlab code by Nima Moshtagh.
Copyright (c) 2009, Nima Moshtagh
}
