% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{occurenceMatrix}
\alias{occurenceMatrix}
\title{Occurence Matrix}
\usage{
occurenceMatrix(values)
}
\arguments{
\item{values}{a list of lists}
}
\description{
Creates an occurence matrix from gene lists (rows - pathways, cols - genes).
}
