\name{pathExplore-package}
\alias{pathExplore-package}
\alias{pathExplore}
\docType{package}
\title{
\packageTitle{pathExplore}
}
\description{
\packageDescription{pathExplore}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{pathExplore}
\packageIndices{pathExplore}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{pathExplore}

Maintainer: \packageMaintainer{pathExplore}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# simple examples of the most important functions
}
