% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{pathwaySimilarity}
\alias{pathwaySimilarity}
\title{Calculates a similarity matrix.}
\usage{
pathwaySimilarity(enrichment, method = "jaccard")
}
\arguments{
\item{enrichment}{a data frame containing enrichment results.}

\item{method}{a method to be used. Available values: 'jaccard', 'cosine', 'mutualInfo'.}
}
\description{
Calculates a similarity matrix of all pathways in the enrichment.
}
