% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkPlot.R
\name{prepareEnrichmentClusters}
\alias{prepareEnrichmentClusters}
\title{Enrichment Clusters}
\usage{
prepareEnrichmentClusters(enrichment, clusters, params)
}
\description{
Prepares enrichment and clustering results for \code{drawPlot} method.
}
