% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{setClusterNames}
\alias{setClusterNames}
\title{Selects best fitting name for a cluster of pathways.}
\usage{
setClusterNames(sim, clusters, method = "pagerank")
}
\arguments{
\item{sim}{similarity matrix used to detect the clusters}

\item{clusters}{a vector of clusters, with names indicating the pathway name}

\item{method}{a method for setting cluster names. Available method include 'pagerank',
'hits', 'wordcloud' and 'none'}
}
\description{
Selects best fitting name for a cluster of pathways.
}
