% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{similarityCorrelation}
\alias{similarityCorrelation}
\title{Calculates correlation between genes in pathways.}
\usage{
similarityCorrelation(genes)
}
\description{
Calculates correlation between genes in pathways.
}
