% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{similarityCosine}
\alias{similarityCosine}
\title{Calculates cosine similarity between all pathways.}
\usage{
similarityCosine(genes)
}
\description{
Calculates cosine similarity between all pathways.
}
