% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{similarityJaccard}
\alias{similarityJaccard}
\title{Calculates jaccard similarity between all pathways.}
\usage{
similarityJaccard(genes)
}
\description{
Calculates jaccard similarity between all pathways.
}
