% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{similarityMutualInformation}
\alias{similarityMutualInformation}
\title{Calculates mutual information between all pathways.}
\usage{
similarityMutualInformation(genes)
}
\description{
Calculates mutual information between all pathways.
}
