% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{validateEnrichment}
\alias{validateEnrichment}
\title{Enrichment Validation}
\usage{
validateEnrichment(enrichment, colorBy = NULL, nodeSize = NULL, verbose = TRUE)
}
\arguments{
\item{enrichment}{data frame containing enrichment results}

\item{colorBy}{which column will be used to color \code{enrichmentNetwork} nodes. Will try to set
automatically if it is NULL}

\item{nodeSize}{which column will be used to set size of \code{enrichmentNetwork} nodes. Will try
to set automatically if it is NULL}

\item{verbose}{enable / disable log messages}
}
\description{
Validates enrichment input for use with \code{enrichmentNetwork} method and checks
for available columns. Currently only validates clusterProfiler results.
}
\seealso{
\code{enrichmentData}
}
