% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkPlot.R
\name{enrichmentNetwork.plot}
\alias{enrichmentNetwork.plot}
\title{Enrichment Network}
\usage{
enrichmentNetwork.plot(
  dt,
  sim,
  clust,
  innerCutoff = 0.1,
  outerCutoff = 0.5,
  colorType = c("nes", "pval"),
  pCutoff = -10,
  drawEllipses = FALSE,
  fontSize = 5,
  repelLabels = FALSE
)
}
\arguments{
\item{dt}{output from \code{enrichmentNetwork.prepareEnrichmentClusters}}

\item{sim}{similarity matrix}

\item{innerCutoff}{similarity cutoff for in-cluster nodes}

\item{outerCutoff}{similarity cutoff for between-cluster nodes}

\item{colorType}{how to colour the nodes: \code{'nes'} - will center around 0 with blue min and
red max, \code{'pval'} - will use log transform on the colorBy column and colour range [-10, 0]}

\item{pCutoff}{adjust p-value colouring cutoff when using \code{colorType = 'pval'}}

\item{drawEllipses}{enable / disable ellipse drawing}
}
\value{
\code{ggplot} object.
}
\description{
Creates a \code{ggplot} object from pathway similarity matrix and clustering results.
}
