% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkPlot.R
\name{enrichmentNetwork}
\alias{enrichmentNetwork}
\title{Enrichment Network}
\usage{
enrichmentNetwork(
  enrichment,
  simMethod = c("jaccard", "cosine", "cor"),
  clustMethod = c("markov", "hier", "spectral"),
  clustNameMethod = c("pagerank", "hits", "none"),
  colorBy = NULL,
  nodeSize = NULL,
  innerCutoff = 0.1,
  outerCutoff = 0.5,
  colorType = c("nes", "pval"),
  pCutoff = -10,
  drawEllipses = FALSE,
  fontSize = 5,
  repelLabels = FALSE,
  minClusterSize = 2,
  plotOnly = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{enrichment}{a data frame containing enrichment results.}

\item{simMethod}{method for calculating similarity matrix between the pathways. Available
methods: \code{'jaccard'}, \code{'cosine'} and \code{'cor'}}

\item{clustMethod}{method for detecting pathway clusters. Available methods: \code{'markov'},
\code{'hier'} and \code{'spectral'}. Using \code{'spectral'} method requires that you have
\code{Spectrum} package installed.}

\item{clustNameMethod}{method for selecting cluster names. Available methods: \code{'pagerank'},
\code{'hits'} and \code{'none'}}

\item{colorBy}{which column should be used to color the nodes. If NULL, will be selected by
default (either NES or p-value will be used)}

\item{nodeSize}{which column should be used to set the node size. If NULL, will be selected by
default (either setSize or Count will be used)}

\item{innerCutoff}{similarity cutoff for in-cluster nodes}

\item{outerCutoff}{similarity cutoff for between-cluster nodes}

\item{colorType}{how to colour the nodes: \code{'nes'} - will center around 0 with blue min and
red max, \code{'pval'} - will use log transform on the colorBy column and adjust color range}

\item{pCutoff}{adjust p-value colouring cutoff when using \code{colorType = 'pval'}}

\item{drawEllipses}{enable / disable ellipse drawing}

\item{fontSize}{adjust cluster label font size}

\item{repelLabels}{should cluster label positions be corrected}

\item{minClusterSize}{min number of nodes in a single cluster}

\item{plotOnly}{default: TRUE. If set to false, returns clustering results as well as the plot
(a list object is returned).}

\item{verbose}{enable / disable log messages}
}
\value{
\code{ggplot} object.
}
\description{
Displays enrichment results as an easy to interpret cluster network.
}
\seealso{
\code{enrichmentData}, \code{validateEnrichment}
}
