% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{findClusters}
\alias{findClusters}
\title{Pathway Clustering}
\usage{
findClusters(
  sim,
  minClusterSize = 2,
  method = c("markov", "hier", "spectral"),
  nameMethod = c("pagerank", "hits", "none"),
  verbose = FALSE
)
}
\arguments{
\item{sim}{similarity matrix}

\item{minClusterSize}{minimum cluster size (integer, >= 1). This parameter is ignored when using
\code{method = 'spectral'}}

\item{method}{a method to be used for detecting clusters. Available methods are: \code{'markov'},
\code{'hier'} and \code{'spectral'}. Using \code{'spectral'} method requires that you have
\code{Spectrum} package installed.}

\item{nameMethod}{a method for setting cluster names. Available methods are: \code{'pagerank'},
\code{'hits'}, and \code{'none'}}

\item{verbose}{enable / disable log messages}
}
\description{
Finds clusters in a similarity matrix.
}
