% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{findClustersHier}
\alias{findClustersHier}
\title{Hierarchical Clustering}
\usage{
findClustersHier(sim, minClusterSize = 2, verbose = FALSE)
}
\arguments{
\item{sim}{similarity matrix}

\item{minClusterSize}{minimum cluster size}

\item{verbose}{enable / disable log messages}
}
\description{
Finds clusters using hierarchical algorithm.
}
