% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{findClustersMarkov}
\alias{findClustersMarkov}
\title{Markov Cluster Algorithm}
\usage{
findClustersMarkov(sim, minClusterSize = 2, verbose = FALSE)
}
\arguments{
\item{sim}{similarity matrix}

\item{minClusterSize}{minimum cluster size}

\item{verbose}{enable / disable log messages}
}
\description{
Finds clusters in a similarity matrix using Markov Cluster Algorithm.
}
