% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{findClustersSpectral}
\alias{findClustersSpectral}
\title{Spectral Clustering}
\usage{
findClustersSpectral(sim, verbose = FALSE)
}
\arguments{
\item{sim}{similarity matrix}

\item{verbose}{enable / disable log messages}
}
\description{
Finds clusters using adaptive spectral clustering. Using this method requires that
you have \code{Spectrum} package installed.
}
