% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{pathwaySimilarity}
\alias{pathwaySimilarity}
\title{Calculates a similarity matrix.}
\usage{
pathwaySimilarity(enrichment, method = "jaccard", useOnlyCore = TRUE)
}
\arguments{
\item{enrichment}{a data table containing enrichment results.}

\item{method}{a method to be used. Available values: 'jaccard', 'cosine', 'mutualInfo'.}

\item{useOnlyCore}{logical value to specify whether to use all genes in the pathway or only
the core enrichment.}
}
\description{
Calculates a similarity matrix.
}
