% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{pathwaySimilarity}
\alias{pathwaySimilarity}
\title{Similarity Matrix}
\usage{
pathwaySimilarity(
  enrichment,
  geneCol,
  pathCol = "Description",
  method = c("jaccard", "cosine", "cor")
)
}
\arguments{
\item{enrichment}{a data frame containing enrichment results}

\item{geneCol}{which column contains gene lists}

\item{pathCol}{which column contains path descriptions (human readable text, not ids!)}

\item{method}{a method to be used. Available values: \code{'jaccard'}, \code{'cosine'} and
\code{'cor'}}
}
\description{
Calculates a similarity matrix of all pathways in the enrichment.
}
