% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{similarityCosine}
\alias{similarityCosine}
\title{Pathway Cosine Similarity}
\usage{
similarityCosine(genes)
}
\arguments{
\item{genes}{genes in all pathways. Should be a list of lists, which each entry containing the
genes that belong to a specific pathway (names of the list should be the pathways)}
}
\description{
Calculates cosine similarity between all pathways.
}
